Final update to number literal handler
Updated my regexps for the conditional statements below the while loop. The regexps  /[\.]{1}/ for double type tokens and /^[0-9]$/ for integer type tokens were not precise in checking if the number literal lexeme is valid or not. They only found one character that was present in the lexeme, but that wasn't good enough.