Initial commit of lexer mod
Wanted to build a better lexer. The old version of the lexer in lexer.js is a bit ineffective in scanning the entire input as it requires that each lexeme to be separated by whitespaces for it to tokenize properly.  The old version is also unwieldy and more complex than it had to be since I crammed all of the novel features I thought would be cool to use into that module.