Multi line comment ignorer fix
The while loop condition before the fix caused the loop to stop skipping lexemes when it reached a newline lexeme before the */ lexeme. Also the line under the while loop divvy[--index] causes an infinite recursion syndrome. After skipping the lexemes, as soon as the ignorer reaches the */ lexeme the loop is supposed to break, but the while loop jumps back to the last ignored lexeme then breaks, then jump back to that lexeme and breaks, over and over and over again.