Revised the punctuator lexeme algorithm
While writing the switch statement of 3 cases and a default case, the lexer combined the characters  )  and  {  together into  ){. That lexeme was invalid which was why the individual tokens for characters ) and { weren't generated. Greatly reduced the punctuator lexeme algorithm. It doesn't make sense to add those individual characters together. Totally missed that.