Squashed a bug
My mistake of having the scanner push the string "<newline>" into the divvy array when it sees the end of a line in streamer.txt could've caused an out of bounds error. My single line comment ignorer algorithm wouldn't stop running beyond the end of the line because it's looking for this string "\n" and can't find one. If the break condition wasn't there, the evaluator algorithm would be 1 index out of the divvy array bounds. I didn't want to update the match condition of the ignorer so I made the scanner push "\n" again when it finds newlines.